set up authentication and authorization for deleting posts

+++ controllers/api/v1/posts_api.js
@@ -23,31 +23,20 @@ module.exports.destroy = async function(req, res) {
         let post = await Post.findById(req.params.id);
 
         // .id means converting the object id into string
-        // if(post.user == req.user.id) {
-            post.remove();
+        if(post.user == req.user.id) {
+            post.remove();  
     
             await Comment.deleteMany({post: req.params.id});
 
-            // if(req.xhr) {
-            //     return res.status(200).json({
-            //         data: {
-            //             post_id: req.params.id
-            //         },
-            //         message: "Post deleted!"
-            //     });
-            // }
-
-            // req.flash('success', 'Post and associated comments deleted!');
-
-            // return res.redirect('back');
             return res.status(200).json({
                 message: "Post and associated comments deleted!"
             });
-        // }
-        // else {
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+        }
+        else {
+            return res.status(401).json({
+                message: "You cannot delete this post!"
+            });
+        }
     } catch (err) {
         // req.flash('error', err); 
         // return res.redirect('back');
